+++ lib/triangulate.mli
@@ -1,0 +1,1 @@
+val triangulate : ?eps:float -> ?idxs:int array -> V.v2 array -> int list list