+++ lib/v.ml
@@ -173,10 +173,11 @@ module type S = sig
 
   (** [clockwise_sign ?eps a b c]
 
-    Returns the rotational ordering (around the z-axis) of the points [a], [b],
-    and [c] as a signed float, [-1.] for clockwise, and [1.] for
-    counter-clockwise. If the points are collinear (not forming a valid triangle,
-    within the tolerance of [eps]), [0.] is returned. *)
+    Returns the rotational ordering (around the z-axis, from the perspective of
+    the origin, looking "up" the z-axis) of the points [a], [b], and [c] as a
+    signed float, [1.] for clockwise, and [-1.] for counter-clockwise. If the
+    points are collinear (not forming a valid triangle, within the tolerance of
+    [eps]), [0.] is returned. *)
   val clockwise_sign : ?eps:float -> t -> t -> t -> float
 
   (** [collinear p1 p2 p3]