+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## v0.1.1
 - Add `Mesh.drop_unused_points`
+- Remove `Mesh.add_face` and `Mesh.add_faces`
+- Add `Path2.triangulate` and `Mesh.triangulate`
 
 ## v0.1.0